requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
scipy>=1.7.0
matplotlib>=3.5.0
seaborn>=0.11.0
jupyter>=1.0.0
imbalanced-learn>=0.9.0
shap>=0.41.0
xgboost>=1.7.0
lightgbm>=3.3.0
pydantic>=1.10.0
pyyaml>=6.0
mlflow>=2.0.0
openpyxl>=3.0.0
